Extract ComputedContingencyElement and ConnectivityBreakAnalysis classes from WoodburyEngine.
