+++ src/test/java/com/powsybl/openloadflow/sensi/DcSensitivityAnalysisTest.java
@@ -15,6 +15,7 @@ import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import com.powsybl.iidm.network.test.PhaseShifterTestCaseFactory;
 import com.powsybl.loadflow.LoadFlowParameters;
+import com.powsybl.openloadflow.dc.ComputedContingencyElement;
 import com.powsybl.openloadflow.dc.WoodburyEngine;
 import com.powsybl.openloadflow.dc.equations.DcEquationType;
 import com.powsybl.openloadflow.dc.equations.DcVariableType;
@@ -1053,7 +1054,7 @@ class DcSensitivityAnalysisTest extends AbstractSensitivityAnalysisTest {
         assertEquals("Too many factors groups 3333333, maximum is 2684 for a system with 100000 equations", e.getMessage());
 
         LfNetwork network = Mockito.mock(LfNetwork.class);
-        List<WoodburyEngine.ComputedContingencyElement> contingencyElements = Mockito.mock(List.class);
+        List<ComputedContingencyElement> contingencyElements = Mockito.mock(List.class);
         Mockito.when(contingencyElements.size()).thenReturn(999999);
         e = assertThrows(PowsyblException.class, () -> WoodburyEngine.initContingencyRhs(network, equationSystem, contingencyElements));
         assertEquals("Too many contingency elements 999999, maximum is 2684 for a system with 100000 equations", e.getMessage());
